Implement `MessageFlags`; Refactor flags placement
