+++ discord/http.py
@@ -372,12 +372,6 @@ class HTTPClient:
         r = Route('PATCH', '/channels/{channel_id}/messages/{message_id}', channel_id=channel_id, message_id=message_id)
         return self.request(r, json=fields)
 
-    def suppress_message_embeds(self, channel_id, message_id, *, suppress):
-        payload = { 'suppress': suppress }
-        r = Route('POST', '/channels/{channel_id}/messages/{message_id}/suppress-embeds',
-                  channel_id=channel_id, message_id=message_id)
-        return self.request(r, json=payload)
-
     def add_reaction(self, channel_id, message_id, emoji):
         r = Route('PUT', '/channels/{channel_id}/messages/{message_id}/reactions/{emoji}/@me',
                   channel_id=channel_id, message_id=message_id, emoji=emoji)
@@ -429,6 +423,10 @@ class HTTPClient:
             params['around'] = around
 
         return self.request(Route('GET', '/channels/{channel_id}/messages', channel_id=channel_id), params=params)
+
+    def publish_message(self, channel_id, message_id):
+        return self.request(Route('POST', '/channels/{channel_id}/messages/{message_id}/crosspost', 
+                                  channel_id=channel_id, message_id=message_id))
 
     def pin_message(self, channel_id, message_id):
         return self.request(Route('PUT', '/channels/{channel_id}/pins/{message_id}',
