+++ docs/api.rst
@@ -2410,6 +2410,12 @@ SystemChannelFlags
 .. autoclass:: SystemChannelFlags
     :members:
 
+MessageFlags
+~~~~~~~~~~~~
+
+.. autoclass:: MessageFlags
+	:members:
+
 
 Exceptions
 ------------
